requirements.txt
amqp==1.4.7
anyjson==0.3.3
beaker==1.7.0
billiard==3.3.0.21
bottle==0.12.8
bottle-session==0.5
celery==3.1.19
Django==1.7.1
Flask==0.10.1
gunicorn==19.3.0
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.8
kombu==3.0.29
MarkupSafe==0.23
python-instagram==1.3.2
pytz==2015.7
redis==2.10.3
simplejson==3.8.0
six==1.10.0
Twisted==15.4.0
Werkzeug==0.10.4
zope.interface==4.1.3
